+++ src/axom/inlet/VerifiableScalar.hpp
@@ -32,7 +32,7 @@ class Field;
  *
  * \brief Basic interface for verifiable scalar values of Inlet primitive type,
  * namely int, double, bool, or std::string - implementations can use this
- * directly (inlet::Field) or forward to all elements of a container (inlet::AggregrateField)
+ * directly (inlet::Field) or forward to all elements of a collection (inlet::AggregrateField)
  * 
  * In practice this interface is used for the Field and AggregateField types.
  * In addition to the ability to mark things as required and use user-defined lambdas