+++ src/axom/inlet/examples/user_defined_type.cpp
@@ -7,7 +7,6 @@
 
 #include <iostream>
 #include <unordered_map>
-#include "CLI11/CLI11.hpp"
 #include "axom/slic/core/SimpleLogger.hpp"
 
 using axom::inlet::FunctionType;
@@ -17,277 +16,118 @@ using axom::sidre::DataStore;
 
 namespace inlet = axom::inlet;
 // _inlet_userdef_simple_start
-struct Mesh
+struct Car
 {
-  std::string filename;
-  int serial_ref_iter;
-  int par_ref_iter;
+  std::string make;
+  std::string color;
+  int seats;
+  int horsepower;
 
-  // Each class should define a static method that adds the fields it
-  // will grab from inlet
-  static void defineSchema(inlet::Table& schema)
+  // A function can be used to define a schema for a particular struct
+  // For convenience, it can be implemented as a static method of the struct
+  static void defineSchema(inlet::Table& car_schema)
   {
-    schema.addString("filename", "Path to mesh file");
-    schema.addInt("serial", "Number of serial refinement iterations");
-    schema.addInt("parallel", "Number of parallel refinement iterations");
+    car_schema.addString("make", "Make of car");
+    car_schema.addString("color", "Color of car").defaultValue("red");
+    car_schema.addInt("seats", "Number of seats").range(2, 7);
+    car_schema.addInt("horsepower", "Amount of horsepower");
   }
 };
 // _inlet_userdef_simple_end
 
 // Additionally, each class should specialize this struct as follows
 // in the global namespace so that Inlet can access it
-/**
- * Example Lua definition:
- * \code{.lua}
- * mesh = {
- *    filename = 'data/square.mesh',
- *    serial = 12,
- *    parallel = 7
- * }
- * \endcode
- */
 // _inlet_userdef_simple_frominlet_start
 template <>
-struct FromInlet<Mesh>
+struct FromInlet<Car>
 {
-  Mesh operator()(const inlet::Proxy& base)
+  Car operator()(const inlet::Proxy& input_data)
   {
-    return {base["filename"], base["serial"], base["parallel"]};
+    Car result;
+    result.make = input_data["make"];
+    result.color = input_data["color"];
+    result.seats = input_data["seats"];
+    result.horsepower = input_data["horsepower"];
+    return result;
   }
 };
 // _inlet_userdef_simple_frominlet_end
 
-struct LinearSolver
-{
-  double rel_tol;
-  double abs_tol;
-  int print_level;
-  int max_iter;
-  double dt;
-  int steps;
-  static void defineSchema(inlet::Table& schema)
-  {
-    schema.addDouble("rel_tol", "Relative convergence criterion");
-    schema.addDouble("abs_tol", "Relative convergence criterion");
-    schema.addInt("print_level", "Logging level for iterative solver");
-    schema.addInt("max_iter", "Maximum iteration count");
-    schema.addDouble("dt", "Time step");
-    schema.addInt("steps", "Number of simulation iterations/frames");
+const std::string input = R"(
+  -- A single car
+  car = {
+      make = "BestCompany",
+      seats = 2,
+      horsepower = 200,
+      color = "blue"
   }
-};
 
-/**
- * Example Lua definition:
- * \code{.lua}
- * solver = {
- *    rel_tol = 1.e-6,
- *    abs_tol = 1.e-12,
- *    print_level = 0,
- *    max_iter = 100,
- *    dt = 1.0,
- *    steps = 1 
- * }
- * \endcode
- */
-template <>
-struct FromInlet<LinearSolver>
-{
-  LinearSolver operator()(const inlet::Proxy& base)
-  {
-    LinearSolver lin_solve;
-    lin_solve.rel_tol = base["rel_tol"];
-    lin_solve.abs_tol = base["abs_tol"];
-    lin_solve.print_level = base["print_level"];
-    lin_solve.max_iter = base["max_iter"];
-    lin_solve.dt = base["dt"];
-    lin_solve.steps = base["steps"];
-    return lin_solve;
-  }
-};
-
-// _inlet_userdef_bc_struct_start
-struct BoundaryCondition
-{
-  std::unordered_map<int, int> attrs;
-  // std::functions are nullable - coef/vec_coef act as a sum type here
-  std::function<double(FunctionType::Vector)> coef;
-  std::function<FunctionType::Vector(FunctionType::Vector)> vec_coef;
-  // _inlet_userdef_bc_struct_end
-  static void defineSchema(inlet::Table& schema)
-  {
-    schema.addIntArray("attrs", "List of boundary attributes");
-    // Inlet does not support sum types, so both options are added to the schema
-    // Supported function parameter/return types are Double and Vector
-    schema.addFunction("vec_coef",
-                       inlet::FunctionTag::Vector,    // Return type
-                       {inlet::FunctionTag::Vector},  // Argument types
-                       "The function representing the BC coefficient");
-    // _inlet_userdef_func_coef_start
-    schema.addFunction("coef",
-                       inlet::FunctionTag::Double,    // Return type
-                       {inlet::FunctionTag::Vector},  // Argument types
-                       "The function representing the BC coefficient");
-    // _inlet_userdef_func_coef_end
-  }
-};
-
-/**
- * Example Lua definition:
- * \code{.lua}
- * bc = {
- *   attrs = {
- *      3, 4, 6, 9
- *   }
- *   coef = function (v)
- *     return v.x * 0.12
- *   end
- * }
- * -- or, for vector coefficients:
- * [8] = {
- *   attrs = { [4] = 14, [8] = 62, [6] = 11},
- *   vec_coef = function (v)
- *     scale = 0.12
- *     return v * scale
- *   end
- * }
- * \endcode
- */
-template <>
-struct FromInlet<BoundaryCondition>
-{
-  BoundaryCondition operator()(const inlet::Proxy& base)
-  {
-    BoundaryCondition bc;
-    bc.attrs = base["attrs"];
-    // _inlet_userdef_bc_struct_retrieve_start
-    if(base.contains("vec_coef"))
+  -- A collection of cars
+  fleet = {
     {
-      bc.vec_coef = base["vec_coef"];
-    }
-    else
+      make = "Globex Corp",
+      seats = 3,
+      horsepower = 155,
+      color = "green"
+    },
     {
-      bc.coef = base["coef"];
+      make = "Initech",
+      seats = 4,
+      horsepower = 370
+    },
+    {
+      make = "Cyberdyne",
+      seats = 2,
+      horsepower = 101,
+      color = "silver"
     }
-    // _inlet_userdef_bc_struct_retrieve_end
-    return bc;
   }
-};
+)";
 
-struct ThermalSolver
-{
-  Mesh mesh;
-  LinearSolver solver;
-  std::unordered_map<std::string, BoundaryCondition> bcs;
-  // defineSchema is intended to be used recursively
-  // Tables are created for subobjects and passed to
-  // subobject defineSchema implementations
-  static void defineSchema(inlet::Table& schema)
-  {
-    // _inlet_userdef_simple_usage_start
-    auto& mesh_schema = schema.addStruct("mesh", "Information about the mesh");
-    Mesh::defineSchema(mesh_schema);
-    // _inlet_userdef_simple_usage_end
-    auto& solver_schema = schema.addStruct(
-      "solver",
-      "Information about the iterative solver used for Ku = f");
-    LinearSolver::defineSchema(solver_schema);
-
-    // _inlet_userdef_array_usage_start
-    // Schema only needs to be defined once, will propagate through to each
-    // element of the array, namely, the subtable at each found index in the input file
-    auto& bc_schema =
-      schema.addStructDictionary("bcs", "List of boundary conditions");
-    BoundaryCondition::defineSchema(bc_schema);
-    // _inlet_userdef_array_usage_end
-  }
-};
-
-/**
- * Example Lua definition:
- * \code{.lua}
- * thermal_solver = {
- *    mesh = {
- *      -- see above FromInlet<Mesh>
- *    },
- *    solver = {
- *      -- see above FromInlet<LinearSolver>
- *    },
- *    bcs = {
- *        ["temperature"] = {
- *          -- see above FromInlet<BoundaryCondition>
- *        },
- *        ["flux"] = {
- *          -- see above FromInlet<BoundaryCondition>
- *        },
- *    }
- * }
- * \endcode
- */
-template <>
-struct FromInlet<ThermalSolver>
-{
-  // This is also implicitly recursive - will call the FromInlet
-  // functions defined for the subobjects
-  ThermalSolver operator()(const inlet::Proxy& base)
-  {
-    return {
-      base["mesh"].get<Mesh>(),
-      base["solver"].get<LinearSolver>(),
-      base["bcs"].get<std::unordered_map<std::string, BoundaryCondition>>()};
-  }
-};
-
-int main(int argc, char** argv)
+int main()
 {
   // Inlet requires a SLIC logger to be initialized to output runtime information
   axom::slic::SimpleLogger logger;
 
-  CLI::App app {"Example of Axom's Inlet component with user-defined types"};
-  std::string inputFileName;
-  auto opt = app.add_option("--file", inputFileName, "Path to input file");
-  opt->check(CLI::ExistingFile);
-
-  CLI11_PARSE(app, argc, argv);
-
   DataStore ds;
   auto lr = std::make_unique<LuaReader>();
-  lr->parseFile(inputFileName);
+  lr->parseString(input);
   Inlet inlet(std::move(lr), ds.getRoot());
 
-  // Create a table off the global table for the thermal_solver object
+  // _inlet_userdef_simple_usage_start
+  // Create a table off the global table for the car object
   // then define its schema
-  auto& thermal_solver_schema =
-    inlet.addStruct("thermal_solver",
-                    "Configuration for a thermal conduction module");
-  ThermalSolver::defineSchema(thermal_solver_schema);
+  auto& car_schema = inlet.addStruct("car", "Vehicle description");
+  Car::defineSchema(car_schema);
+  // _inlet_userdef_simple_usage_end
+
+  // _inlet_userdef_collection_usage_start
+  // Create a fleet of cars with the same Car::defineSchema
+  auto& fleet_schema = inlet.addStructArray("fleet", "A collection of cars");
+  Car::defineSchema(fleet_schema);
+  // _inlet_userdef_collection_usage_end
 
   if(!inlet.verify())
   {
     SLIC_ERROR("Inlet failed to verify against provided schema");
   }
 
-  // Read all the data into a thermal solver object
-  auto thermal_solver = inlet["thermal_solver"].get<ThermalSolver>();
+  // Extract the car object
+  Car car = inlet["car"].get<Car>();
+  std::cout << "Info on single car:" << std::endl;
+  std::cout << "make = " << car.make << std::endl;
+  std::cout << "color = " << car.color << std::endl;
+  std::cout << "seats = " << car.seats << std::endl;
+  std::cout << "horsepower = " << car.horsepower << std::endl;
 
-  const FunctionType::Vector vec {1, 2, 3};
-  for(const auto& bc_entry : thermal_solver.bcs)
+  // Extract the cars in the fleet
+  auto fleet = inlet["fleet"].get<std::vector<Car>>();
+  std::cout << "\nInfo on cars in fleet:" << std::endl;
+  for(const Car& car : fleet)
   {
-    const auto& bc = bc_entry.second;
-    if(bc.coef)
-    {
-      auto result = bc.coef(vec);
-      SLIC_INFO(fmt::format("Calling {0} with {1} returned: {2}",
-                            bc_entry.first,
-                            vec,
-                            result));
-    }
-    else if(bc.vec_coef)
-    {
-      auto result = bc.vec_coef(vec);
-      SLIC_INFO(fmt::format("Calling {0} with {1} returned: {2}",
-                            bc_entry.first,
-                            vec,
-                            result));
-    }
+    std::cout << "make = " << car.make << std::endl;
+    std::cout << "color = " << car.color << std::endl;
+    std::cout << "seats = " << car.seats << std::endl;
+    std::cout << "horsepower = " << car.horsepower << std::endl;
   }
 }