+++ src/axom/sidre/examples/sidre_mfem_datacollection_restart.cpp
@@ -198,15 +198,18 @@ int main(int argc, char* argv[])
   dc.Save("sidre_mfem_datacoll_restart_ex", "sidre_hdf5");
 
   // Sample time parameters
-  int n_iter = 10;
-  double dt = 0.05;
+  const int n_iter = 10;
+  const int n_checkpoint = 5;
+  const double dt = 0.05;
 
   for(int i = 0; i < n_iter; i++)
   {
     sim_state.step(dt);
-
-    // then save it at each iteration
-    dc.Save("sidre_mfem_datacoll_restart_ex", "sidre_hdf5");
+    if(i % n_checkpoint == 0)
+    {
+      // then save it at each checkpoint
+      dc.Save("sidre_mfem_datacoll_restart_ex", "sidre_hdf5");
+    }
   }
 
 #if defined(AXOM_USE_MPI) && defined(MFEM_USE_MPI)