+++ RELEASE-NOTES.md
@@ -65,6 +65,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
   datastore is `Load` ed in
 - Inlet: Exposed primal::Vector in Lua for use in input-file-defined functions
 - Inlet: Cleaned up `Table` interface to eliminate ambiguity and duplicated functionality
+- Inlet: Renamed `DocWriter` to `Writer` and refactored its interface
 - Inlet deserialization logic uses ``Proxy`` instead of ``Table`` to support conversions from
   primitives to user-defined types (e.g., string to enum)
 
@@ -75,6 +76,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Inlet: Apply lambda verifiers on generic containers to individual elements
   for consistency
 - Inlet: Fixed a bug relating to nested table lookups of primitive arrays and functions
+- Fixed a bug relating to deeply nested callback functions in Inlet
 
 
 ## [Version 0.4.0] - Release date 2020-09-22